fix(remove-text-loader): 🐛 remove loader from list view
